fix: prevent cloud-init status from erroring
For reasons I can't understand, the cloud-init works but returns an error
status. It says "Failed pickling datasource to /var/lib/cloud/instance/obj.pkl"
whether there is anything in the userdata or not. In the end, I've removed
the error check so that we can look for whether the "status" == "done"